Implement larger group linking
When calculating shifts, check if a group is a subset of some other group. If so, link it to the larger group instead of processing separately.
For example, take a character talking for five seconds. During this time, a one-second sign appears. We can link this sign to the character line to get better shift.
Greatly improves fbf typesetting.