Properly treat audio values as signed
Unsigned values don't work right with small negative data. For example silence in range [-1, 0] becomes total garbage with values [255, 0] when unsigned. This breaks search for anything containing mostly silence.
Instead we're properly normalizing the audio to 0.0 - 1.0. This fixes the silence issue but leads to worse general search in CR audio.