Implement partial error recovery on increasing shift
When grouping is enabled and Sushi encounters more than --rewind-thresh [5] events with considerably different shift, it assumes for this to be an increasing shift error, increases window size (up to --max-window [30]) and tries to process all broken events again.

This is just a partial implementation of the full algorithm. Namely it works only if there's indeed a matching audio segment and it's possible to find it with larger window. Sushi still fails to recover when there's a missing audio segment, samples are needed for that to get implemented.